requirements.txt
absl-py==0.8.0
aiofiles==0.7.0
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.0
arcade==1.3.7
asgiref==3.4.1
astor==0.8.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.1.0
beautifulsoup4==4.10.0
bleach==2.1.4
blis==0.7.4
boto3==1.19.3
botocore==1.22.3
cachetools==4.2.4
catalogue==1.0.0
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
cloudpickle==2.0.0
conllu==4.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.4
decorator==4.3.0
deepface==0.0.68
defusedxml==0.5.0
entity-embed==0.0.6
entrypoints==0.2.3
falcon==3.0.1
filelock==3.0.12
Flask==2.0.2
Flask-Sockets==0.2.1
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
future==0.17.1
gast==0.3.1
gdown==4.2.0
gevent==1.4.0
gevent-websocket==0.10.1
glfw==1.7.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.42.0
gym==0.21.0
-e git+https://github.com/NamanB/VITPAL.git@ef72d537a00b7a4166d86163b20700eac4ef14b5#egg=gym_minigrid&subdirectory=gym-minigrid
h11==0.12.0
h5py==2.10.0
html5lib==1.0.1
idna==2.7
imageio==2.4.1
importlib-metadata==4.8.1
iniconfig==1.1.1
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==2.0.1
jedi==0.12.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
jsonlines==2.0.0
jsonnet==0.17.0
jsonpickle==2.0.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.2
lmdb==1.2.1
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.4.3
mistune==0.8.3
more-itertools==8.10.0
mtcnn==0.1.1
multidict==5.2.0
murmurhash==1.0.5
n2==0.1.7
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.6.2
notebook==5.7.0
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
ordered-set==4.0.2
overrides==3.1.0
packaging==21.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pathy==0.6.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==8.4.0
plac==1.1.3
pluggy==1.0.0
preshed==3.0.5
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.9.1
ptyprocess==0.6.0
py==1.10.0
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.19
pydantic==1.8.2
pygame==1.9.4
pyglet==1.3.2
Pygments==2.2.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
pytorch-lightning==1.2.10
pytorch-metric-learning==0.9.99
pytz==2018.9
PyYAML==6.0
pyzmq==17.1.2
qtconsole==4.4.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.8
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
sentencepiece==0.1.96
simplegeneric==0.8.1
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==2.3.5
spacy-legacy==3.0.8
srsly==1.0.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.3.1
thinc==7.4.5
tokenizers==0.10.3
toml==0.10.2
torch==1.10.0
torchtext==0.9.0
torchvision==0.11.1
tornado==5.1.1
tqdm==4.62.3
traitlets==4.3.2
transformers==4.3.3
typer==0.3.2
typing-extensions==3.10.0.2
Unidecode==1.3.2
urllib3==1.24
uvicorn==0.15.0
wasabi==0.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.4.2
word2number==1.1
wrapt==1.11.2
yarl==1.7.2
zipp==3.6.0
zxing==0.13.1
